Adds template merging.
If a template is created in the same directory as other template files,
the contents of a template will be added and merged in with the existing
templates.

Also, it is possible to pass a comma-separated string (1,2,3,15) in the
template select field to parse through multiple templates at once.

Also, as always, improvements to the Makefile:

* Automatically generate go representations of static assets (template
  files).
* Any changes to template files causes Make to rebuild assets.go files.